ci: Edit workflows
Edit the workflow structure. There are new files `ci.yml` and `ci-publish.yml` which call the existing workflows. As such, the previously existing workflows have been edited to be callable.

The CI workflow runs on any pull request and is also called by the CI and Publish workflow.

The CI and Publish workflow calls the CI workflow and then calls the Publish Python distribution to PyPI and TestPyPI workflow if the CI workflow is successful.