feat: Remove --appimage-extract-and-run option
The original purpose of the --appimage-extract-and-run option was to provide an option for the sake of compatibility. Upon further testing, it was found that the difference in resource usage was so small, it would be better to just have the option enabled by default for appimagetool.

Thus, the option in PyDeployment was removed entirely, and an environment variable, namely `APPIMAGE_EXTRACT_AND_RUN` is set to enabled for the appimagetool command issued as part of the build process for Linux.

In the future, there may be an option to disable this extract and run option, but that will be addressed in a separate commit should the need be demonstrated.